Refactor: Separate create and read functionality tests into own classes
